requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
colorama==0.4.6
coverage==7.13.3
crispy-tailwind==1.0.3
Django==6.0.1
django-allauth==65.14.0
django-browser-reload==1.21.0
django-crispy-forms==2.5
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-watchfiles==1.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
pastel==0.2.1
pluggy==1.6.0
poethepoet==0.40.0
psycopg2-binary==2.9.11
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pylance==2.0.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.37.0
rich==14.3.2
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
watchfiles==1.1.1
